feat: add redux, restcountries api and validation selec countries
