+++ package.json
@@ -21,10 +21,12 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^5.2.1",
+    "@reduxjs/toolkit": "^2.8.2",
     "@tailwindcss/vite": "^4.1.12",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-hook-form": "^7.62.0",
+    "react-redux": "^9.2.0",
     "tailwindcss": "^4.1.12",
     "zod": "^4.0.17"
   },
