+++ src/store/reducers/root-reducer.ts
@@ -1,0 +1,6 @@
+import { combineReducers } from '@reduxjs/toolkit';
+import { restcountries } from '../../utils/restcountries.api';
+
+export const rootReducer = combineReducers({
+  [restcountries.reducerPath]: restcountries.reducer,
+});
