+++ src/types/api.type.ts
@@ -1,0 +1,12 @@
+export type Country = {
+  name: {
+    common: string;
+    official: string;
+    nativeName?: {
+      [languageCode: string]: {
+        official: string;
+        common: string;
+      };
+    };
+  };
+};
